Revert "Merge branch 'develop' into feature/on-boarding"
This reverts commit 585c54fff101153e040ea29a63e210ac0a8b0bd8, reversing
changes made to 6bd5f62e06d2a0e650b977562cd609ea9d7c3ef2.
