+++ .env
@@ -1,1 +1,2 @@
 NEXT_PUBLIC_GRAPHQL_ENDPOINT=https://staging.a-comosus.com/graphql/
+GRAPHQL_ENDPOINT=https://staging.a-comosus.com/graphql/
