+++ .env.development
@@ -1,0 +1,2 @@
+NEXT_PUBLIC_GRAPHQL_ENDPOINT=http://localhost:3100/graphql/
+GRAPHQL_ENDPOINT=http://localhost:3100/graphql/
