+++ CI-CD/Jenkinsfile
@@ -2,8 +2,6 @@ pipeline {
     agent any
 
     options {
-        // This is required if you want to clean before build
-        skipDefaultCheckout(true)
         //  Specifying how many artifacts and logs we want to persist. 
         //  In this case we are going to persist the last 10 builds.
         buildDiscarder(logRotator(numToKeepStr: '10'))
@@ -36,10 +34,6 @@ pipeline {
             steps {
                 echo 'Cleanup docker space'
                 sh 'docker system prune -f'
-                 // Clean before build
-                cleanWs()
-                // We need to explicitly checkout from SCM here
-                checkout scm
             }
         }
 
@@ -53,39 +47,38 @@ pipeline {
         stage('Install dependencies') {
             steps {
                 echo 'Install dependencies...'
-                sh 'npm install'
+                sh 'yarn install'
             }
         }
 
         stage('Test') {
             steps {
                 echo 'Testing..'
-                sh 'npm run codegen' 
-                sh 'npm run test:coverage'
+                sh 'yarn test:coverage'
             }
         }
         
         stage('Checking for linter error') {
             steps {
                 echo 'Checking for linter error....'
-                sh 'npm run lint'
+                sh 'yarn lint'
             }
         }
-        /*
+
         stage('Build') {
             steps {
                 echo 'Building....'
-                sh 'npm run build'
+                sh 'yarn build:with-codegen'
             }
         }
 
         stage ('Integrated Test') {
             steps {
                 echo 'Integrated Test....'
-                sh 'npm run e2e:start-server:record'
+                sh 'yarn e2e:start-server:record'
             }
         }
-       */
+       
 
         // stage ('Container Integrated Test') {
         //     steps {
@@ -106,7 +99,7 @@ pipeline {
             steps {
                 echo 'build image with ecr tage...'
                 withCredentials([string(credentialsId: 'AWS_REPOSITORY_URL_SECRET', variable: 'AWS_ECR_URL')]) {
-                    sh "docker build --build-arg N_G_ENDPOINT=${N_G_ENDPOINT} -t ${AWS_ECR_URL}:${IMAGE_VERSION} -f Dockerfile.Alex ."
+                    sh "docker build --build-arg N_G_ENDPOINT=${N_G_ENDPOINT} --build-arg G_ENDPOINT=${G_ENDPOINT} -t ${AWS_ECR_URL}:${IMAGE_VERSION} ."
                 }
 
                 echo 'upload to ECR'
@@ -163,7 +156,9 @@ pipeline {
 
     post {
         always {
-            deleteDir()
+            withCredentials([string(credentialsId: 'AWS_REPOSITORY_URL_SECRET', variable: 'AWS_ECR_URL')]) {
+                deleteDir()
+            }
         }
     }
 }
