+++ codegen.yml
@@ -1,5 +1,5 @@
 overwrite: true
-schema: ${NEXT_PUBLIC_GRAPHQL_ENDPOINT}
+schema: ${GRAPHQL_ENDPOINT}
 
 documents: './src/**/*.gql'
 generates:
