+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "codegen": "graphql-codegen --config codegen.yml -r dotenv-expand/config",
+    "codegen": "graphql-codegen --config codegen.yml -r dotenv/config",
     "dev": "next dev",
     "build": "next build",
     "build:with-codegen": "npm run codegen && npm run build",
@@ -25,6 +25,11 @@
     "@fontsource/montserrat": "^4.5.10",
     "@hookform/resolvers": "^2.8.10",
     "axios": "^0.27.2",
+    "bootstrap": "^5.1.3",
+    "dotenv": "^16.0.1",
+    "framer-motion": "^6",
+    "graphql": "^16.5.0",
+    "graphql-request": "^4.2.0",
     "i18next": "^21.8.3",
     "jquery": "^3.6.0",
     "lodash": "^4.17.21",
@@ -59,8 +64,6 @@
     "@typescript-eslint/eslint-plugin": "^5.26.0",
     "@typescript-eslint/parser": "^5.26.0",
     "cypress": "^10.0.2",
-    "dotenv": "^16.0.1",
-    "dotenv-expand": "^8.0.3",
     "eslint": "8.15.0",
     "eslint-config-next": "12.1.6",
     "eslint-config-prettier": "^8.5.0",
@@ -68,7 +71,6 @@
     "jest": "^28.1.0",
     "jest-environment-jsdom": "^28.1.0",
     "prettier": "^2.6.2",
-    "sass": "^1.52.3",
     "start-server-and-test": "^1.14.0",
     "ts-jest": "^28.0.2",
     "typescript": "4.6.4"
