--- public/locales/cn/admin.json
@@ -1,16 +1,0 @@
-{
-  "admin": {
-    "title": "凤梨酥 🍍 管理员面板"
-  },
-  "link": {
-    "editor": {
-      "menu": {
-        "add-new-link": "添加新链接",
-        "explore": "探索更多选择"
-      }
-    },
-    "preview": {
-      "share": "分享你的凤梨酥吧!"
-    }
-  }
-}
