+++ public/locales/cn/auth.json
@@ -56,9 +56,6 @@
         "required": "请同意我们的政策"
       }
     },
-    "success": {
-      "message": "欢迎加入！"
-    },
     "error": {
       "user-exist": "用户已存在，请换一个用户名"
     }
@@ -107,7 +104,5 @@
        "Tech":"🖥第一生产力",
        "Travel":"✈️诗和远方"
     }
-  "logout": {
-    "message": "下次再见啦~"
   }
 }
