--- public/locales/en/admin.json
@@ -1,16 +1,0 @@
-{
-  "admin": {
-    "title": "A-Comosus 🍍 Admin"
-  },
-  "link": {
-    "editor": {
-      "menu": {
-        "add-new-link": "Add New Link",
-        "explore": "Explore"
-      }
-    },
-    "preview": {
-      "share": "Share your A-Comosus"
-    }
-  }
-}
