+++ public/locales/en/auth.json
@@ -55,9 +55,6 @@
       "description": "Don't have an account? ",
       "action": "Create One"
     },
-    "success": {
-      "message": "Welcome Aboard!"
-    },
     "error": {
       "user-exist": "User already exist, please use another name."
     }
@@ -105,7 +102,5 @@
        "Tech":"🖥Tech",
        "Travel":"✈️Travel & Tourism"
     }
-  "logout": {
-    "message": "See you around!"
   }
 }
