+++ public/locales/en/common.json
@@ -1,7 +1,4 @@
 {
-  "site": {
-    "title": "A-Comosus 🍍"
-  },
   "button": {
     "disabled": "Disabled"
   },
@@ -14,16 +11,5 @@
   },
   "login": {
     "title": "Just Login"
-  },
-  "sidebar": {
-    "link": "Links",
-    "appearance": "Appearance",
-    "settings": "Settings",
-    "analytics": "Analytics",
-    "upgrade": "Upgrade",
-    "support": "Support",
-    "new": "What's new",
-    "my-account": "my-account",
-    "logout": "logout"
   }
 }
