+++ src/common/components/forms/FormErrorMessage.tsx
@@ -1,16 +1,15 @@
 import React from 'react';
 
 import { MdError } from 'react-icons/md';
-import { FormErrorMessage, FormErrorMessageProps } from '@chakra-ui/react';
+import { FormErrorMessage } from '@chakra-ui/react';
 
 type CustomFormErrorMessageProps = {
   testId: string;
   error?: string;
-} & FormErrorMessageProps;
+};
 export default function CustomFormErrorMessage({
   testId,
   error,
-  ...props
 }: CustomFormErrorMessageProps) {
   return (
     <FormErrorMessage
@@ -18,7 +17,6 @@ export default function CustomFormErrorMessage({
       flex={1}
       alignSelf="flex-start"
       gap="5px"
-      {...props}
     >
       <MdError />
       {error ?? 'Undefined'}
