+++ src/common/components/index.ts
@@ -1,5 +1,5 @@
 // Container Components
-export { default as PageContainer } from './containers/PageContainer';
+export { default as AppContainer } from './containers/AppContainer';
 
 // Generic Components
 export { default as Logo } from './generics/Logo';
@@ -7,8 +7,6 @@ export { default as NavBar } from './generics/NavBar';
 export { default as Footer } from './generics/Footer';
 export { default as SelectLanguage } from './generics/SelectLanguage';
 export { default as ToggleThemeButton } from './generics/ToggleThemeButton';
-export { default as Sidebar } from './generics/Sidebar/Sidebar';
-
 // Form Components
 export { default as Button } from './forms/Button';
 export { default as Input } from './forms/Input';
