+++ src/common/contexts/ApiClient.tsx
@@ -1,10 +1,11 @@
 import React, { useMemo } from 'react';
-import { GraphQLClient } from 'graphql-request';
-import axios, { AxiosInstance } from 'axios';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { ReactQueryDevtools } from 'react-query/devtools';
+import { GraphQLClient } from 'graphql-request';
+import axios, { AxiosInstance } from 'axios';
 
-import { useAuth } from '@common/contexts';
+import useStorage from '@src/utils/hooks/useStorage';
+import { AuthKey, StorageType } from '@src/constants/StorageKey';
 
 const apiEndpoint = {
   gql:
@@ -16,12 +17,11 @@ const apiEndpoint = {
     process.env.DEFAULT_ENDPOINT ??
     'http://localhost:3100/graphql/',
 };
-
 type ApiClientContextType = {
   gqlClient: GraphQLClient;
   restClient: AxiosInstance;
-  queryClient: QueryClient;
 };
+
 const ApiClientContext = React.createContext<ApiClientContextType>(
   {} as ApiClientContextType,
 );
@@ -29,9 +29,16 @@ const ApiClientContext = React.createContext<ApiClientContextType>(
 type ApiClientProviderProps = {
   children: React.ReactNode;
 };
+
+export function useApiClient() {
+  return React.useContext(ApiClientContext);
+}
+
 export function ApiClientProvider({ children }: ApiClientProviderProps) {
   const { accessToken } = useAuth();
 
+  // Common header
+  // AuthOption 1: Insert Auth Token here with axios.interceptors.request
   const headers = useMemo(() => {
     const _header = {
       'Content-Type': 'application/json',
@@ -49,6 +56,9 @@ export function ApiClientProvider({ children }: ApiClientProviderProps) {
     const client = new GraphQLClient(apiEndpoint.gql, {
       headers,
     });
+
+    // AuthOption 2: Insert Auth Token here with axios.interceptors.request
+
     return client;
   }, [headers]);
 
@@ -58,23 +68,23 @@ export function ApiClientProvider({ children }: ApiClientProviderProps) {
       baseURL: apiEndpoint.rest,
       headers,
     });
+
+    // AuthOption 2: Insert Auth Token here with axios.interceptors.request
+
     return client;
   }, [headers]);
 
-  const queryStaleTime = parseInt(
-    process.env.NEXT_PUBLIC_QUERY_STALE_TIME_IN_MINUTES ?? '900',
-    10,
-  );
+  const queryStaleTime = parseInt(process.env.QUERY_STALE_TIME ?? '1', 10);
   const queryClient = new QueryClient({
     defaultOptions: {
       queries: {
-        staleTime: queryStaleTime * 60 * 1000,
+        staleTime: queryStaleTime * 1000,
       },
     },
   });
 
   return (
-    <ApiClientContext.Provider value={{ gqlClient, restClient, queryClient }}>
+    <ApiClientContext.Provider value={{ gqlClient, restClient }}>
       <QueryClientProvider client={queryClient}>
         <ReactQueryDevtools initialIsOpen={false} />
         {children}
@@ -83,6 +93,10 @@ export function ApiClientProvider({ children }: ApiClientProviderProps) {
   );
 }
 
-export function useApiClient() {
-  return React.useContext(ApiClientContext);
+export function useAuth() {
+  const [accessToken, setAccessToken] = useStorage(
+    AuthKey.AccessToken,
+    StorageType.Session,
+  );
+  return { accessToken, setAccessToken };
 }
