--- src/common/contexts/apiClientContext/server.gql
@@ -1,5 +1,0 @@
-query GetServerInfo {
-  server {
-    status
-  }
-}
