+++ src/common/contexts/index.ts
@@ -1,7 +1,2 @@
+export { ApiClientProvider, useApiClient, useAuth } from './ApiClient';
 export { ThemeContextProvider } from './themeContext/ThemeProvider';
-export {
-  ApiClientProvider,
-  useApiClient,
-} from './apiClientContext/ApiClientContext';
-export { AuthProvider, useAuth } from './AuthContext';
-export { UserProvider, useUser } from './userContext/UserContext';
