--- src/common/contexts/userContext/user.gql
@@ -1,7 +1,0 @@
-query FindUserById($payload: FindByIdInput!) {
-  findUserById(data: $payload) {
-    id
-    username
-    email
-  }
-}
