+++ src/constants/PageRoutes.ts
@@ -1,7 +1,3 @@
-export enum GlobalRoute {
-  Root = '/',
-}
-
 export enum AuthRoute {
   SignUp = '/sign-up',
   Login = '/login',
