--- src/constants/QueryKey.ts
@@ -1,4 +1,0 @@
-export enum LinkQueries {
-  FindAllOfAnUser = 'FindLinksOfUserByUserId',
-  FindByUsername = 'FindUserByUsername',
-}
