--- src/modules/admin/admin.d.ts
@@ -1,13 +1,0 @@
-type EditableLinkFormType = {
-  isVisible: boolean;
-  title: string;
-  url: string;
-};
-
-// Best to have this managed by backend
-type Link = {
-  id: string;
-  title: string;
-  url: string;
-  logoUrl?: string | null;
-};
