--- src/modules/admin/admin.gql
@@ -1,42 +1,0 @@
-mutation CreateNewLink($payload: CreateLinkInput!) {
-  createLink(data: $payload) {
-    id
-  }
-}
-
-query FindLinksOfUserByUserId($payload: FindLinksOfUserByUserIdInput!) {
-  findLinksOfUserByUserId(data: $payload) {
-    id
-    isDraft
-    isVisible
-    title
-    url
-  }
-}
-
-mutation UpdateLink($payload: UpdateLinkInput!) {
-  updateLink(data: $payload) {
-    id
-    isVisible
-    title
-    url
-  }
-}
-
-mutation DeleteLinkById($payload: DeleteLinkInput!) {
-  deleteLinkById(data: $payload)
-}
-
-query FindUserByUsername($payload: FindByUsernameArgs!) {
-  findUserByUsername(username: $payload) {
-    id
-    email
-    username
-    links {
-      id
-      title
-      url
-      logoUrl
-    }
-  }
-}
