--- src/modules/admin/components/index.ts
@@ -1,1 +1,0 @@
-export { default as AppContainer } from './AppContainer';
