--- src/modules/admin/sections/index.ts
@@ -1,2 +1,0 @@
-export { default as LinkEditor } from './LinkEditor/LinkEditor';
-export { default as ProfilePreview } from './ProfilePreview/ProfilePreview';
