+++ src/pages/_app.tsx
@@ -4,21 +4,15 @@ import { ni18nConfig } from '../ni18n.config';
 import '@coreui/coreui/dist/css/coreui.min.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
 
-import {
-  ApiClientProvider,
-  AuthProvider,
-  ThemeContextProvider,
-} from '@common/contexts';
+import { ApiClientProvider, ThemeContextProvider } from '@common/contexts';
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
-    <AuthProvider>
-      <ApiClientProvider>
-        <ThemeContextProvider>
-          <Component {...pageProps} />
-        </ThemeContextProvider>
-      </ApiClientProvider>
-    </AuthProvider>
+    <ApiClientProvider>
+      <ThemeContextProvider>
+        <Component {...pageProps} />
+      </ThemeContextProvider>
+    </ApiClientProvider>
   );
 }
