+++ src/pages/forget-password.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next';
 import { useApiClient } from '@common/contexts';
 
 import { HStack, VStack } from '@chakra-ui/react';
-import { PageContainer, Logo } from '@src/common/components';
+import { AppContainer, Logo } from '@src/common/components';
 import { ForgetPasswordForm } from '@src/modules/auth';
 import { isNil } from 'lodash';
 import { useForgetPasswordMutation } from '@generated/graphql.queries';
@@ -35,7 +35,7 @@ export default function ForgetPassword() {
   };
 
   return (
-    <PageContainer head={head}>
+    <AppContainer head={head}>
       <VStack flex={1} padding="1rem">
         <HStack width="100%">
           <Logo height="5rem" />
@@ -47,6 +47,6 @@ export default function ForgetPassword() {
           successMessage={successMessage}
         />
       </VStack>
-    </PageContainer>
+    </AppContainer>
   );
 }
