--- src/types.d.ts
@@ -1,11 +1,0 @@
-type Head = {
-  title: string;
-  metas?: {
-    name: string;
-    content: string;
-  }[];
-  links?: {
-    rel: string;
-    href: string;
-  }[];
-};
